BRAIN_TUMOR_APP/requirements.txt
streamlit
tensorflow
numpy
opencv-python
pillow
matplotlib
scikit-learn
gdown
fpdf
requests
